Pull method out of hosted payment control and place in configuration to be reusable. Add ObsidianComponent extension and required methods, use the reusable method in both the original place and new location for Obsidian.
